Add reference spectrum to equalization plots and more plotting enhancements
Use  a header  format for  polynomial coefficients  that includes  the
valid  X  range  in  scaled  terms  and  a  count  of  the  number  of
coefficients.

Use double  precision consistently  for polynomial  coefficients. This
includes formatting with sufficient DPs when writing to files.

Many changes to the equalization plots, more to come.

Add  error   handling  for   reading  coefficient,  plot   and  filter
files.  This  includes  being  backward  compatible  for  old  format
refspec.dat files with no header.

git-svn-id: svn+ssh://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@7578 ab8295b8-cf94-4d9e-aec4-7959e3be5d79

+++ PhaseEqualizationDialog.hpp
@@ -17,11 +17,11 @@ class PhaseEqualizationDialog
 public:
   explicit PhaseEqualizationDialog (QSettings *
                                     , QDir const& data_directory
-                                    , QVector<float> const& coefficients
+                                    , QVector<double> const& coefficients
                                     , QWidget * = nullptr);
   Q_SLOT void show ();
 
-  Q_SIGNAL void phase_equalization_changed (QVector<float> const&);
+  Q_SIGNAL void phase_equalization_changed (QVector<double> const&);
 
 private:
   class impl;
